# src/modules/dashboard/dto/update-dashboard.dto.ts
import { IsString, IsOptional, IsBoolean, IsArray, ValidateNested, IsNumber } from 'class-validator';
import { Type } from 'class-transformer';
import { CreateDashboardWidgetDto } from './create-dashboard.dto';

export class UpdateDashboardDto {
  @IsOptional()
  @IsString()
  name?: string;

  @IsOptional()
  @IsBoolean()
  isDefault?: boolean;

  @IsOptional()
  @IsArray()
  @ValidateNested({ each: true })
  @Type(() => CreateDashboardWidgetDto)
  widgets?: CreateDashboardWidgetDto[];
}