# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/dashboards": {
      "get": {
        "operationId": "DashboardController_findForUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Dashboards"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "DashboardController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDashboardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Dashboards"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/dashboards/defaults": {
      "get": {
        "operationId": "DashboardController_findDefaults",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Dashboards"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/dashboards/{id}": {
      "put": {
        "operationId": "DashboardController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDashboardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Dashboards"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "DashboardController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Dashboards"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/dashboards/seed-defaults": {
      "post": {
        "operationId": "DashboardController_seedDefaults",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Dashboards"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "IGreen HRMS API",
    "description": "Enterprise Human Resource Management System API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDashboardWidgetDto": {
        "type": "object",
        "properties": {
          "widgetType": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "colSpan": {
            "type": "number"
          },
          "rowSpan": {
            "type": "number"
          },
          "config": {
            "type": "object"
          }
        },
        "required": [
          "widgetType",
          "title",
          "position"
        ]
      },
      "CreateDashboardDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "roleName": {
            "type": "string"
          },
          "widgets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateDashboardWidgetDto"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateDashboardDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "widgets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateDashboardWidgetDto"
            }
          }
        }
      }
    }
  }
}